Move heavy e2e to new folder
