+++ integration/e2e/deployments/config-all-in-one-azurite.yaml
