+++ integration/e2e/deployments/config-all-in-one-local.yaml
