+++ integration/e2e/deployments/config-all-in-one-s3.yaml
