+++ integration/e2e/deployments/config-scalable-single-binary.yaml
