+++ integration/e2e/encodings_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 	"time"
 
+	util2 "github.com/grafana/tempo/integration/util"
 	v2 "github.com/grafana/tempo/tempodb/encoding/v2"
 
 	"github.com/grafana/e2e"
@@ -13,10 +14,10 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/grafana/tempo/cmd/tempo/app"
-	"github.com/grafana/tempo/integration"
 	"github.com/grafana/tempo/integration/e2e/backend"
+	"github.com/grafana/tempo/integration/util"
 	"github.com/grafana/tempo/pkg/httpclient"
-	"github.com/grafana/tempo/pkg/util"
+	tempoUtil "github.com/grafana/tempo/pkg/util"
 	"github.com/grafana/tempo/tempodb/encoding"
 )
 
@@ -35,7 +36,7 @@ func TestEncodings(t *testing.T) {
 
 			// copy config template to shared directory and expand template variables
 			tmplConfig := map[string]any{"Version": enc.Version()}
-			config, err := integration.CopyTemplateToSharedDir(s, configAllEncodings, "config.yaml", tmplConfig)
+			config, err := util2.CopyTemplateToSharedDir(s, configAllEncodings, "config.yaml", tmplConfig)
 			require.NoError(t, err)
 
 			// load final config
@@ -49,44 +50,44 @@ func TestEncodings(t *testing.T) {
 			_, err = backend.New(s, cfg)
 			require.NoError(t, err)
 
-			tempo := integration.NewTempoAllInOne()
+			tempo := util2.NewTempoAllInOne()
 			require.NoError(t, s.StartAndWaitReady(tempo))
 
 			// Get port for the Jaeger gRPC receiver endpoint
-			c, err := integration.NewJaegerGRPCClient(tempo.Endpoint(14250))
+			c, err := util2.NewJaegerGRPCClient(tempo.Endpoint(14250))
 			require.NoError(t, err)
 			require.NotNil(t, c)
 
-			info := util.NewTraceInfo(time.Now(), "")
+			info := tempoUtil.NewTraceInfo(time.Now(), "")
 			require.NoError(t, info.EmitAllBatches(c))
 
 			expected, err := info.ConstructTraceFromEpoch()
 			require.NoError(t, err)
 
 			// test metrics
-			require.NoError(t, tempo.WaitSumMetrics(e2e.Equals(spanCount(expected)), "tempo_distributor_spans_received_total"))
+			require.NoError(t, tempo.WaitSumMetrics(e2e.Equals(util.SpanCount(expected)), "tempo_distributor_spans_received_total"))
 
 			// test echo
-			assertEcho(t, "http://"+tempo.Endpoint(3200)+"/api/echo")
+			util.AssertEcho(t, "http://"+tempo.Endpoint(3200)+"/api/echo")
 
 			apiClient := httpclient.New("http://"+tempo.Endpoint(3200), "")
 
 			// query an in-memory trace
-			queryAndAssertTrace(t, apiClient, info)
+			util.QueryAndAssertTrace(t, apiClient, info)
 
 			// wait trace_idle_time and ensure trace is created in ingester
 			require.NoError(t, tempo.WaitSumMetricsWithOptions(e2e.Less(3), []string{"tempo_ingester_traces_created_total"}, e2e.WaitMissingMetrics))
 
 			// flush trace to backend
-			callFlush(t, tempo)
+			util.CallFlush(t, tempo)
 
 			// v2 does not support querying and must be skipped
 			if enc.Version() != v2.VersionString {
 				// search for trace in backend multiple times with different attributes to make sure
 				// we search with different scopes and with attributes from dedicated columns
 				for i := 0; i < repeatedSearchCount; i++ {
-					integration.SearchAndAssertTrace(t, apiClient, info)
-					integration.SearchTraceQLAndAssertTrace(t, apiClient, info)
+					util2.SearchAndAssertTrace(t, apiClient, info)
+					util2.SearchTraceQLAndAssertTrace(t, apiClient, info)
 				}
 			}
 
@@ -94,7 +95,7 @@ func TestEncodings(t *testing.T) {
 			time.Sleep(10 * time.Second)
 
 			// force clear completed block
-			callFlush(t, tempo)
+			util.CallFlush(t, tempo)
 
 			// test metrics
 			require.NoError(t, tempo.WaitSumMetrics(e2e.Equals(1), "tempo_ingester_blocks_flushed_total"))
@@ -104,10 +105,10 @@ func TestEncodings(t *testing.T) {
 			}
 
 			// query trace - should fetch from backend
-			queryAndAssertTrace(t, apiClient, info)
+			util.QueryAndAssertTrace(t, apiClient, info)
 
 			// create grpc client used for streaming
-			grpcClient, err := integration.NewSearchGRPCClient(context.Background(), tempo.Endpoint(3200))
+			grpcClient, err := util2.NewSearchGRPCClient(context.Background(), tempo.Endpoint(3200))
 			require.NoError(t, err)
 
 			if enc.Version() == v2.VersionString {
@@ -119,9 +120,9 @@ func TestEncodings(t *testing.T) {
 			now := time.Now()
 			for i := 0; i < repeatedSearchCount; i++ {
 				// search the backend. this works b/c we're passing a start/end AND setting query ingesters within min/max to 0
-				integration.SearchAndAssertTraceBackend(t, apiClient, info, now.Add(-20*time.Minute).Unix(), now.Unix())
+				util2.SearchAndAssertTraceBackend(t, apiClient, info, now.Add(-20*time.Minute).Unix(), now.Unix())
 				// find the trace with streaming. using the http server b/c that's what Grafana will do
-				integration.SearchStreamAndAssertTrace(t, context.Background(), grpcClient, info, now.Add(-20*time.Minute).Unix(), now.Unix())
+				util2.SearchStreamAndAssertTrace(t, context.Background(), grpcClient, info, now.Add(-20*time.Minute).Unix(), now.Unix())
 			}
 		})
 	}
