+++ integration/e2e/serverless/serverless_test.go
@@ -4,13 +4,13 @@ import (
 	"testing"
 	"time"
 
+	"github.com/grafana/tempo/integration/util"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/e2e"
 	e2e_db "github.com/grafana/e2e/db"
 
-	util "github.com/grafana/tempo/integration"
 	"github.com/grafana/tempo/pkg/httpclient"
 	tempoUtil "github.com/grafana/tempo/pkg/util"
 	"github.com/grafana/tempo/tempodb/backend"
